requirements.txt
pillow
PyOpenGL
PyOpenGL 
PyOpenGL_accelerate
pygame
imgui[pygame]